fix testcase json decode

+++ wasmtest/contracts-rust/helloworld/src/lib.rs
@@ -6,7 +6,7 @@ use ostd::prelude::*;
 use ostd::runtime;
 
 #[no_mangle]
-pub fn add(a: u32, b: u32) -> u32 {
+pub fn add(a: u64, b: u64) -> u64 {
     a + b
 }
